fix: knowing bug when uploading none attachment

+++ backend/src/controllers/ticketController.js
@@ -13,17 +13,21 @@ const ticketController = {
   async createTicket(req, res, next) {
     try {
       const { file } = req;
-      const fileSizeInMB = (file?.size || 0) / 1024 / 1024;
-      const fileType = file?.mimetype.split("/")[0];
 
-      if (fileSizeInMB > 2) {
-        await fs.promises.unlink(file.path);
-        return ResponseAPI.error(res, "File size exceeds 2MB", 400);
+      if (file) {
+        const fileSizeInMB = (file.size || 0) / 1024 / 1024;
+        if (fileSizeInMB > 2) {
+          await fs.promises.unlink(file.path);
+          return ResponseAPI.error(res, "File size exceeds 2MB", 400);
+        }
       }
 
-      if (fileType !== "image") {
-        await fs.promises.unlink(file.path);
-        return ResponseAPI.error(res, "File must be an image", 400);
+      if (file) {
+        const fileType = file.mimetype.split("/")[0];
+        if (fileType !== "image") {
+          await fs.promises.unlink(file.path);
+          return ResponseAPI.error(res, "File must be an image", 400);
+        }
       }
 
       const ticket = await Ticket.create({
